dashboards/requirements.txt
streamlit
pandas
plotly
redshift-connector
sqlalchemy
psycopg2-binary
python-dotenv